Removed Access Tokens page and rename the AccessTokensPage to AccessTokens
